chore: turns spinner into a partial
