+++ app/views/layouts/_spinner.html.erb
@@ -1,0 +1,5 @@
+<%= stylesheet_link_tag 'spinner.css' %>
+
+<div id="loading-spinner">
+  <div class="spinner"></div>
+</div>
