+++ app/views/layouts/application.html.erb
@@ -15,9 +15,7 @@
   </head>
 
   <body>
-    <div id="loading-spinner">
-      <div class="spinner"></div>
-    </div>
+    <%= render 'layouts/spinner' %>
 
     <div class="p-8">
       <%= yield %>
